Tidy up custom transition classes
